Add save-doc.js.
